third_party/blink/web_tests/platform/mac-mac14/external/wpt/css/css-scrollbars/scrollbar-width-001-expected.txt

This is a testharness.js-based test.
[FAIL] horizontal-tb, ltr, scrollbar-width auto
  assert_less_than: ltr auto scrollWidth expected a number less than 200 but got 200
[FAIL] horizontal-tb, ltr, scrollbar-width thin
  assert_less_than: ltr thin scrollWidth expected a number less than 200 but got 200
[FAIL] horizontal-tb, rtl, scrollbar-width auto
  assert_less_than: rtl auto scrollWidth expected a number less than 200 but got 200
[FAIL] horizontal-tb, rtl, scrollbar-width thin
  assert_less_than: rtl thin scrollWidth expected a number less than 200 but got 200
Harness: the test ran to completion.
